Scenario change employee

+++ test/dtu/planning/acceptance_tests/GenerateReportSteps.java
@@ -36,7 +36,7 @@ public class GenerateReportSteps {
 	public void theActivityIsEstimatedToLastHours(Integer hours) throws ActivityNotFoundException, OperationNotAllowedException {
 		Project project = planningAppHolder.getPlanningApp().searchForProject(projectHolder.getProject().getProjectNumber());
 		Activity activity = project.getActivityByName(activityHolder.getActivity().getName());
-		double hours1 = new Double(hours); 
+		float hours1 = new Float(hours);  
 		activity.setExpectedAmountOfHours(hours1);
 	}
 	
@@ -62,7 +62,7 @@ public class GenerateReportSteps {
 	
 	@Then("A report over the project is generated with {int} hours estimated on activity with name {string}")
 	public void aReportOverTheProjectIsGeneratedWithHoursEstimatedOnActivityWithName(Integer hours, String activityName) throws ActivityNotFoundException {
-		double hours1 = new Double(hours); 
+		float hours1 = new Float(hours); 
 		// Third argument in assert equals with doubble is delta
 		assertEquals(report.getEstimatedTimeForActivity(activityName),hours1,0);
 	}
@@ -70,7 +70,7 @@ public class GenerateReportSteps {
 	
 	@Then("A report over the project is generated with {int} hours reported on activity with name {string}")
 	public void aReportOverTheProjectIsGeneratedWithHoursReportedOnActivityWithName(Integer hours, String activityName) throws ActivityNotFoundException{
-		double hours1 = new Double(hours); 
+		float hours1 = new Float(hours); 
 		// Third argument in assert equals with doubble is delta
 		assertEquals(report.getReportedTimeForActivity(activityName),hours1,0);	
 	}
